requirements.txt
Pillow==9.3.0
customtkinter~=5.0.3
future~=0.18.2